# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/options": {
            "get": {
                "tags": [
                    "Опции"
                ],
                "summary": "Получить все доступные опции",
                "operationId": "aba522e39edcbd0a845ebb6cf802a499",
                "responses": {
                    "200": {
                        "description": "Список всех опций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Option"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Опции"
                ],
                "summary": "Создать новую опцию",
                "operationId": "321d8300ff36fc5056c2f6f0132602cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название опции",
                                        "type": "string",
                                        "example": "Климат-контроль"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Опция успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/options/{id}": {
            "get": {
                "tags": [
                    "Опции"
                ],
                "summary": "Получить информацию об опции по ID",
                "operationId": "98465114297e0e658b4250d937997608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об опции",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Опции"
                ],
                "summary": "Удалить опцию",
                "operationId": "75240c34e59673b44a9e49f58fe805f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Опция успешно удалена"
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Опции"
                ],
                "summary": "Обновить информацию об опции",
                "operationId": "51b131a0e07f1ac57537736cae1afc81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название опции",
                                        "type": "string",
                                        "example": "Климат-контроль"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Опция успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Option": {
                "title": "Опция",
                "description": "Опция, для конфигурации автомобиля",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор опции",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название опции",
                        "type": "string",
                        "example": "Климат-контроль"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T12:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}